refactor(enterprise/coderd/connectionlog): merge batcher into DBBackend, never drop logs
- Moved batcher logic from coderd/connectionlogbatcher into the existing
  enterprise/coderd/connectionlog package, replacing the separate dbBackend
  and batchDBBackend with a single DBBackend that batches writes.
- Changed Add() to Upsert() that blocks when the buffer is full instead of
  dropping entries, ensuring no connection logs are ever lost.
- Increased default batch size from 500 to 1000.
- Added Close() to ConnectionLogger that closes all io.Closer backends.
- Wired shutdown in enterprise/coderd/coderd.go to flush on close.
- Deleted coderd/connectionlogbatcher/ package.
